backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.77.0
anyio==4.12.1
appdirs==1.4.4
argcomplete==3.6.3
asgiref==3.11.1
async-timeout==5.0.1
asyncer==0.0.12
asyncpg==0.31.0
attr==0.3.1
attrs==25.4.0
azure-core==1.38.0
azure-storage-blob==12.28.0
banks==2.2.0
beautifulsoup4==4.14.3
bidict==0.23.1
black==26.1.0
bleach==5.0.1
boto3==1.42.45
botocore==1.42.45
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
chainlit==2.9.5
charset-normalizer==3.4.4
chevron==0.14.0
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
croniter==6.0.0
cryptography==46.0.3
cuid==0.4
cycler==0.12.1
dataclasses-json==0.6.7
datamodel-code-generator==0.26.1
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
Django==5.1.15
django-annoying==0.10.6
django-cors-headers==4.7.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.2.3
django-filter==24.3
django-migration-linter==5.2.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rq==3.1
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
docstring_parser==0.17.0
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
drf-spectacular==0.28.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1
expiringdict==1.2.2
Faker==40.4.0
fastapi==0.128.0
filelock==3.20.3
filetype==1.2.0
flupy==1.2.3
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.1.0
genson==1.3.0
google-api-core==2.29.0
google-auth==2.48.0
google-cloud-appengine-logging==1.8.0
google-cloud-audit-log==0.4.0
google-cloud-core==2.5.0
google-cloud-logging==3.13.0
google-cloud-storage==2.19.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
griffe==1.15.0
grpc-google-iam-v1==0.14.3
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.3.5
hyperframe==6.1.0
idna==3.11
ijson==3.4.0.post0
importlib_metadata==8.7.1
inflect==5.6.2
inflection==0.5.1
isodate==0.7.2
isort==5.13.2
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.1.0
joblib==1.5.3
jsf==0.11.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
label-studio==1.22.0
label-studio-sdk==2.0.16
labelImg==1.8.6
launchdarkly-server-sdk==8.2.1
Lazify==0.4.0
literalai==0.1.201
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.12
llama-index-cli==0.5.3
llama-index-core==0.14.12
llama-index-embeddings-openai==0.5.1
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.2
llama-index-llms-llama-cpp==0.5.1
llama-index-llms-openai==0.6.12
llama-index-readers-file==0.5.6
llama-index-readers-llama-parse==0.5.1
llama-index-workflows==2.11.7
llama-parse==0.6.54
llama_cpp_python==0.3.16
lockfile==0.12.2
lxml==6.0.2
lxml_html_clean==0.4.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mcp==1.26.0
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.2
numpy==2.2.6
openai==1.109.1
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-agno==0.51.1
opentelemetry-instrumentation-alephalpha==0.51.1
opentelemetry-instrumentation-anthropic==0.51.1
opentelemetry-instrumentation-bedrock==0.51.1
opentelemetry-instrumentation-chromadb==0.51.1
opentelemetry-instrumentation-cohere==0.51.1
opentelemetry-instrumentation-crewai==0.51.1
opentelemetry-instrumentation-google-generativeai==0.51.1
opentelemetry-instrumentation-groq==0.51.1
opentelemetry-instrumentation-haystack==0.51.1
opentelemetry-instrumentation-lancedb==0.51.1
opentelemetry-instrumentation-langchain==0.51.1
opentelemetry-instrumentation-llamaindex==0.51.1
opentelemetry-instrumentation-logging==0.60b1
opentelemetry-instrumentation-marqo==0.51.1
opentelemetry-instrumentation-mcp==0.51.1
opentelemetry-instrumentation-milvus==0.51.1
opentelemetry-instrumentation-mistralai==0.51.1
opentelemetry-instrumentation-ollama==0.51.1
opentelemetry-instrumentation-openai==0.51.1
opentelemetry-instrumentation-openai-agents==0.51.1
opentelemetry-instrumentation-pinecone==0.51.1
opentelemetry-instrumentation-qdrant==0.51.1
opentelemetry-instrumentation-redis==0.60b1
opentelemetry-instrumentation-replicate==0.51.1
opentelemetry-instrumentation-requests==0.60b1
opentelemetry-instrumentation-sagemaker==0.51.1
opentelemetry-instrumentation-sqlalchemy==0.60b1
opentelemetry-instrumentation-threading==0.60b1
opentelemetry-instrumentation-together==0.51.1
opentelemetry-instrumentation-transformers==0.51.1
opentelemetry-instrumentation-urllib3==0.60b1
opentelemetry-instrumentation-vertexai==0.51.1
opentelemetry-instrumentation-watsonx==0.51.1
opentelemetry-instrumentation-weaviate==0.51.1
opentelemetry-instrumentation-writer==0.51.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-semantic-conventions-ai==0.4.13
opentelemetry-util-http==0.60b1
ordered-set==4.0.2
packaging==25.0
pandas==2.3.3
pathspec==1.0.4
pgvector==0.3.6
pillow==12.1.0
platformdirs==4.5.1
polars==1.38.1
polars-runtime-32==1.38.1
postgrest==2.27.2
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.4
psutil==7.2.2
psycopg==3.3.2
psycopg-binary==3.3.2
psycopg2-binary==2.9.11
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyboxen==1.3.0
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.10.1
pyparsing==3.3.1
pypdf==6.6.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.18.0
pyRFC3339==2.1.0
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.13.0
python-jose==3.5.0
python-json-logger==2.0.4
python-multipart==0.0.21
python-socketio==5.16.0
pytokens==0.4.1
pytz==2022.7.1
pywin32==311
PyYAML==6.0.3
realtime==2.27.2
redis==5.2.1
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-file==3.0.1
requests-mock==1.12.1
rich==14.2.0
rpds-py==0.30.0
rq==2.6.1
rsa==4.9.1
rstr==3.2.2
rules==3.4
s3transfer==0.16.0
safetensors==0.7.0
scipy==1.15.3
semver==3.0.4
sentry-sdk==2.52.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smart_open==7.5.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.1
SQLAlchemy==2.0.45
sqlparse==0.5.5
sse-starlette==3.2.0
starlette==0.50.0
storage3==2.27.2
StrEnum==0.4.15
strictyaml==1.7.3
striprtf==0.0.26
supabase==2.27.2
supabase-auth==2.27.2
supabase-functions==2.27.2
sympy==1.14.0
syncer==2.0.3
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tldextract==5.3.1
tokenizers==0.22.2
toml==0.10.2
tomli==2.4.0
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.1
traceloop-sdk==0.51.1
typer-slim==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ua-parser==1.0.1
ua-parser-builtins==202602
ujson==5.11.0
ultralytics==8.4.13
ultralytics-thop==2.0.18
uritemplate==4.2.0
urllib3==2.6.3
user-agents==2.2.0
uuid_utils==0.14.0
uvicorn==0.40.0
vecs==0.4.5
watchfiles==1.1.1
webencodings==0.5.1
websockets==15.0.1
wrapt==1.17.3
wsproto==1.3.2
xmljson==0.2.1
yarl==1.22.0
zipp==3.23.0
